requirements.txt
numpy
streamlit
pandas
pycaret>=3.3.2
openpyxl
xlsxwriter
python-dotenv
gdown